+++ test/data/dataset/dataset_test.py
@@ -4,9 +4,9 @@ import torch
 
 from howl.data.common.example import AudioClipExample
 from howl.data.common.metadata import AudioClipMetadata
+from howl.data.common.searcher import WordTranscriptSearcher
 from howl.data.common.vocab import Vocab
 from howl.data.dataset.dataset import AudioDataset, DatasetType
-from howl.data.searcher import WordTranscriptSearcher
 from howl.settings import SETTINGS
