+++ test/model/inference_test.py
@@ -1,7 +1,7 @@
 import unittest
 
+from howl.data.common.searcher import WordTranscriptSearcher
 from howl.data.common.vocab import Vocab
-from howl.data.searcher import WordTranscriptSearcher
 from howl.settings import SETTINGS
