+++ training/run/stitch_vocab_samples.py
@@ -1,9 +1,9 @@
 import argparse
 from pathlib import Path
 
+from howl.data.common.labeler import WordFrameLabeler
 from howl.data.common.vocab import Vocab
 from howl.data.dataset import AudioDatasetWriter, WakeWordDatasetLoader
-from howl.data.labeler import WordFrameLabeler
 from howl.data.stitcher import WordStitcher
 from howl.settings import SETTINGS
